% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stan_model.R
\name{summary_stan_model}
\alias{summary_stan_model}
\title{Produce a summary of a Stan model for paper}
\usage{
summary_stan_model(mod, nfixefs = NULL)
}
\arguments{
\item{mod}{A Stan model.}

\item{nfixefs}{Number of so-called fixed effects.}
}
\value{
A dataframe with
\itemize{
  \item \strong{rows} containing intercept and other fixed effects
  \item \strong{columns} containing comparison, mean, lower, upper,
  probability that effect b < 0
}
}
\description{
Produce a summary of a Stan model for paper
}
\examples{
# See SafaviEtAl2016-vignette.Rmd for complete examples
stanm1_tab<-summary_stan_model(stanm1,nfixefs=3)
}
\seealso{
\url{http://mc-stan.org}.
}

